backend-new/pyproject.toml
[project]
name = "hack-nu"
version = "0.1.0"
description = "Evaluates solutions on a given dataset"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
    "rich>=13.7.0",
    "aiohttp>=3.9.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "ipykernel>=7.2.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.9.0",
    "openai>=1.0.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "accelerate>=0.30.0",
]
